Crate ruma_html

Opinionated HTML parsing and manipulating library.

Like the rest of the Ruma crates, this crate is primarily meant to be used for the Matrix protocol. It should be able to be used to interact with any HTML document but will offer APIs focused on specificities of HTML in the Matrix specification..

Structs

ElementData

The data of an HTML element.

Html

An HTML fragment.

Node

An HTML node.

SanitizerConfig

Configuration to sanitize HTML tags and attributes.

Enums

HtmlSanitizerMode

What HTML tags and attributes should be kept by the sanitizer.

RemoveReplyFallback

Whether to remove the rich reply fallback while sanitizing.

Functions

remove_html_reply_fallback

Remove the rich reply fallback of the given HTML string.

sanitize_html

Sanitize the given HTML string.