Struct ruma_html::Html

pub struct Html { /* private fields */ }

An HTML fragment.

To get the serialized HTML, use its Display implementation. Due to the fact that the HTML is parsed, note that malformed HTML and comments will be stripped from the output.

Implementations

impl Html

pub fn parse(string: &str) -> Self

Construct a new Html by parsing the given string.

This is infallible, any error encountered while parsing the HTML is logged with tracing::debug!.

pub fn sanitize(&mut self)

Sanitize this HTML according to the Matrix specification.

This is equivalent to calling Self::sanitize_with() with a config value of SanitizerConfig::compat().remove_reply_fallback().

pub fn sanitize_with(&mut self, config: SanitizerConfig)

Sanitize this HTML according to the given configuration.